requirements.txt
numpy<2.0
pandas
matplotlib
seaborn
plotly
scikit-learn
scikit-surprise
streamlit
pyarrow
openpyxl
wordcloud